nanobody_antigen/requirements.txt
numpy
pandas
scikit-learn
matplotlib
biopython